requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
pydantic
chromadb
requests
python-dotenv
typing-extensions
mlflow
psycopg2-binary  
